+++ samples/ContactSync/ContactSync.NewApp/ContactSync.NewApp.Database/Migrations/20230627-174137-create-new-schema.sql
@@ -1,0 +1,4 @@
+-- Create schema: [New]
+
+CREATE SCHEMA [New]
+    AUTHORIZATION [dbo];